Added Radio and Timing to Base
Implemented the radio code from BaseStationRadioTest into the main base station function. Tested and working. Implemented millis() based timing to the three test types. Tested and functional. Established a search for the kill switch command. Tested and functional. The actual sensor input and SD card writing still needs to be implemented, but we have a stand that will recieve commands from the remote and send commands to the ESC.